1st iteration of refactoring is done: presenter, convert
