+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,0 +1,30 @@
+import UIKit
+
+final class MovieQuizPresenter {
+    
+    let questionAmount = 10
+    private var currentQuestionIndex = 0
+    
+    func isLastQuestion() -> Bool {
+        currentQuestionIndex == questionAmount - 1
+    }
+    
+    func resetQuestionIndex() {
+        currentQuestionIndex = 0
+    }
+    
+    func switchToNextQuestion() {
+        currentQuestionIndex += 1
+    }
+    
+    func convert(model: QuizQuestion) -> QuizStepViewModel {
+        let result = QuizStepViewModel(
+            image: UIImage(data: model.image) ?? UIImage(),
+            question: model.text,
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionAmount)")
+        return result
+    }
+    
+    
+    
+}
